+++ src/bookmark_reader/chrome.rs
@@ -4,11 +4,9 @@ use log::{debug, trace};
 use serde_json::{Map, Value};
 use std::io::Read;
 
-/// Bookmark reader to read bookmarks from Chromium or Google Chrome.
-#[derive(Clone, Copy)]
-pub struct ChromeBookmarkReader;
+pub struct Chrome;
 
-impl ChromeBookmarkReader {
+impl Chrome {
     fn select_bookmark(obj: &Map<String, Value>, bookmarks: &mut SourceBookmarks) {
         trace!("json object: {obj:#?}");
 
@@ -84,20 +82,20 @@ impl ChromeBookmarkReader {
     }
 }
 
+/// Bookmark reader to read bookmarks from Chromium or Google Chrome.
+#[derive(Clone, Copy)]
+pub struct ChromeBookmarkReader;
+
 impl ReadBookmark for ChromeBookmarkReader {
     fn name(&self) -> &'static str {
         "Chrome/Chromium"
     }
 
     fn extension(&self) -> Option<&str> {
-        None
+        Some("json")
     }
 
-    fn select(
-        &self,
-        reader: &mut dyn Read,
-    ) -> Result<Option<Box<dyn ReadBookmark>>, anyhow::Error> {
-        let raw_bookmarks = self.read(reader)?;
+    fn select(&self, raw_bookmarks: &str) -> Result<Option<Box<dyn ReadBookmark>>, anyhow::Error> {
         let value: Value = serde_json::from_str(&raw_bookmarks)?;
 
         match value {
@@ -130,7 +128,57 @@ impl ReadBookmark for ChromeBookmarkReader {
     ) -> Result<(), anyhow::Error> {
         debug!("Parse bookmarks from {}", self.name());
         let value: Value = serde_json::from_str(raw_bookmarks)?;
-        Self::traverse_json(&value, bookmarks, source);
+        Chrome::traverse_json(&value, bookmarks, source);
+        Ok(())
+    }
+}
+
+#[derive(Clone, Copy)]
+pub struct ChromeNoExtensionBookmarkReader;
+
+impl ReadBookmark for ChromeNoExtensionBookmarkReader {
+    fn name(&self) -> &'static str {
+        "Chrome/Chromium (no extension)"
+    }
+
+    fn extension(&self) -> Option<&str> {
+        None
+    }
+
+    fn select(&self, raw_bookmarks: &str) -> Result<Option<Box<dyn ReadBookmark>>, anyhow::Error> {
+        let value: Value = serde_json::from_str(&raw_bookmarks)?;
+
+        match value {
+            Value::Object(obj) => {
+                if obj.get("checksum").is_some()
+                    && obj.get("roots").is_some()
+                    && obj.get("version").is_some()
+                {
+                    Ok(Some(Box::new(*self)))
+                } else {
+                    Ok(None)
+                }
+            }
+            _ => Ok(None),
+        }
+    }
+
+    fn read(&self, reader: &mut dyn Read) -> Result<String, anyhow::Error> {
+        debug!("Read bookmarks from {}", self.name());
+        let mut bookmarks = Vec::new();
+        reader.read_to_end(&mut bookmarks)?;
+        Ok(String::from_utf8(bookmarks)?)
+    }
+
+    fn parse(
+        &self,
+        raw_bookmarks: &str,
+        source: &Source,
+        bookmarks: &mut SourceBookmarks,
+    ) -> Result<(), anyhow::Error> {
+        debug!("Parse bookmarks from {}", self.name());
+        let value: Value = serde_json::from_str(raw_bookmarks)?;
+        Chrome::traverse_json(&value, bookmarks, source);
         Ok(())
     }
 }
@@ -142,7 +190,7 @@ mod tests {
 
     #[test]
     fn test_parse_all() {
-        let source_path = Path::new("test_data/source/bookmarks_google-chrome.json");
+        let source_path = Path::new("test_data/source/bookmarks_chrome.json");
         assert!(source_path.exists());
 
         let bookmark_reader = ChromeBookmarkReader;
@@ -164,11 +212,58 @@ mod tests {
     }
 
     #[test]
+    fn test_parse_all_no_extension() {
+        let source_path = Path::new("test_data/source/bookmarks_chrome_no_extension");
+        assert!(source_path.exists());
+
+        let bookmark_reader = ChromeNoExtensionBookmarkReader;
+        let mut bookmark_file = bookmark_reader.open(source_path).unwrap();
+
+        let bookmarks = bookmark_reader.read(&mut bookmark_file).unwrap();
+        let mut source_bookmarks = SourceBookmarks::new();
+        let source = Source::new(source_path, vec![]);
+
+        let res = bookmark_reader.parse(&bookmarks, &source, &mut source_bookmarks);
+        assert!(res.is_ok(), "{}", res.unwrap_err());
+
+        assert_eq!(source_bookmarks.bookmarks, HashSet::from_iter([
+            String::from("https://www.deepl.com/translator"),
+            String::from("https://www.quantamagazine.org/how-mathematical-curves-power-cryptography-20220919/"),
+            String::from("https://en.wikipedia.org/wiki/Design_Patterns"),
+            String::from("https://doc.rust-lang.org/book/title-page.html"),
+        ]));
+    }
+
+    #[test]
     fn test_parse_folder() {
-        let source_path = Path::new("test_data/source/bookmarks_google-chrome.json");
+        let source_path = Path::new("test_data/source/bookmarks_chrome.json");
         assert!(source_path.exists());
 
         let bookmark_reader = ChromeBookmarkReader;
+        let mut bookmark_file = bookmark_reader.open(source_path).unwrap();
+
+        let bookmarks = bookmark_reader.read(&mut bookmark_file).unwrap();
+        let mut source_bookmarks = SourceBookmarks::new();
+        let source = Source::new(source_path, vec![String::from("dev")]);
+
+        let res = bookmark_reader.parse(&bookmarks, &source, &mut source_bookmarks);
+        assert!(res.is_ok(), "{}", res.unwrap_err());
+
+        assert_eq!(
+            source_bookmarks.bookmarks,
+            HashSet::from_iter([
+                String::from("https://en.wikipedia.org/wiki/Design_Patterns"),
+                String::from("https://doc.rust-lang.org/book/title-page.html"),
+            ])
+        );
+    }
+
+    #[test]
+    fn test_parse_folder_no_extension() {
+        let source_path = Path::new("test_data/source/bookmarks_chrome_no_extension");
+        assert!(source_path.exists());
+
+        let bookmark_reader = ChromeNoExtensionBookmarkReader;
         let mut bookmark_file = bookmark_reader.open(source_path).unwrap();
 
         let bookmarks = bookmark_reader.read(&mut bookmark_file).unwrap();
