+++ src/bookmarks/source_bookmarks.rs
@@ -47,53 +47,53 @@ mod tests {
 
     #[test]
     fn test_read_firefox() {
+        let bookmark_path = Path::new("test_data/source/bookmarks_firefox.json");
+        let bookmark_readers = BookmarkReaders::new();
+        let source = Source::new(bookmark_path, vec![]);
+        let source_reader = SourceReader::new(&source, &bookmark_readers.0).unwrap();
+        assert_eq!(source_reader.bookmark_reader().name(), "Firefox");
+    }
+
+    #[test]
+    fn test_read_firefox_compressed() {
         let bookmark_path = Path::new("test_data/source/bookmarks_firefox.jsonlz4");
         test_utils::create_compressed_bookmarks(bookmark_path);
         let bookmark_readers = BookmarkReaders::new();
         let source = Source::new(bookmark_path, vec![]);
         let source_reader = SourceReader::new(&source, &bookmark_readers.0).unwrap();
-        let res = SourceBookmarks::read(&mut [source_reader]);
-        assert!(res.is_ok(), "{}", res.unwrap_err());
-        let source_bookmarks = res.unwrap();
-        assert_eq!(source_bookmarks.bookmarks, HashSet::from_iter([
-            String::from("https://www.mozilla.org/en-US/firefox/central/"),
-            String::from("https://www.quantamagazine.org/how-mathematical-curves-power-cryptography-20220919/"),
-            String::from("https://en.wikipedia.org/wiki/Design_Patterns"),
-            String::from("https://doc.rust-lang.org/book/title-page.html")
-        ]));
+        assert_eq!(
+            source_reader.bookmark_reader().name(),
+            "Firefox (compressed)"
+        );
     }
 
     #[test]
     fn test_read_chrome() {
-        let bookmark_path = Path::new("test_data/source/bookmarks_google-chrome.json");
+        let bookmark_path = Path::new("test_data/source/bookmarks_chrome.json");
         let bookmark_readers = BookmarkReaders::new();
         let source = Source::new(bookmark_path, vec![]);
         let source_reader = SourceReader::new(&source, &bookmark_readers.0).unwrap();
-        let res = SourceBookmarks::read(&mut [source_reader]);
-        assert!(res.is_ok(), "{}", res.unwrap_err());
-        let source_bookmarks = res.unwrap();
-        assert_eq!(source_bookmarks.bookmarks, HashSet::from_iter([
-            String::from("https://www.deepl.com/translator"),
-            String::from("https://www.quantamagazine.org/how-mathematical-curves-power-cryptography-20220919/"),
-            String::from("https://en.wikipedia.org/wiki/Design_Patterns"),
-            String::from("https://doc.rust-lang.org/book/title-page.html"),
-        ]));
+        assert_eq!(source_reader.bookmark_reader().name(), "Chrome/Chromium");
+    }
+
+    #[test]
+    fn test_read_chrome_no_extension() {
+        let bookmark_path = Path::new("test_data/source/bookmarks_chrome_no_extension");
+        let bookmark_readers = BookmarkReaders::new();
+        let source = Source::new(bookmark_path, vec![]);
+        let source_reader = SourceReader::new(&source, &bookmark_readers.0).unwrap();
+        assert_eq!(
+            source_reader.bookmark_reader().name(),
+            "Chrome/Chromium (no extension)"
+        );
     }
 
     #[test]
     fn test_read_simple() {
         let bookmark_path = Path::new("test_data/source/bookmarks_simple.txt");
         let bookmark_readers = BookmarkReaders::new();
-
         let source = Source::new(bookmark_path, vec![]);
         let source_reader = SourceReader::new(&source, &bookmark_readers.0).unwrap();
-        let res = SourceBookmarks::read(&mut [source_reader]);
-        assert!(res.is_ok(), "{}", res.unwrap_err());
-        let source_bookmarks = res.unwrap();
-        assert_eq!(source_bookmarks.bookmarks, HashSet::from_iter([
-            String::from("https://www.quantamagazine.org/how-mathematical-curves-power-cryptography-20220919/"),
-            String::from("https://www.quantamagazine.org/how-galois-groups-used-polynomial-symmetries-to-reshape-math-20210803/"),
-            String::from("https://www.quantamagazine.org/computing-expert-says-programmers-need-more-math-20220517/"),
-        ]));
+        assert_eq!(source_reader.bookmark_reader().name(), "simple");
     }
 }
