+++ src/cmd/update.rs
@@ -85,7 +85,7 @@ mod tests {
         let now = Utc::now();
         let client = MockClient::new();
         let cache = MockCache::new();
-        let bookmark_path = Path::new("test_data/source/bookmarks_google-chrome.json");
+        let bookmark_path = Path::new("test_data/source/bookmarks_chrome.json");
         let source = Source::new(bookmark_path, vec![]);
         let bookmark_readers = BookmarkReaders::new();
         let source_reader = SourceReader::new(&source, &bookmark_readers.0).unwrap();
